Flask Container/requirements.txt
flask
flask-httpauth
azure-iot-hub